backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
openai-agents>=0.6.1
qdrant-client>=1.7.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.3.0
sqlalchemy>=2.0.0
tiktoken>=0.5.0
python-frontmatter>=1.0.0
